회원가입 api 연동

+++ app/src/main/java/com/example/heysrealprojcet/ui/sign_up/age/SignUpAgeViewModel.kt
@@ -13,7 +13,7 @@ class SignUpAgeViewModel : ViewModel() {
    init {
       viewModelScope.launch {
          age.asFlow().collect {
-            UserPreference.age = it
+            UserPreference.age = it.toInt()
             isCorrect()
          }
       }
